Add ability to self image_self_link in stemcell.MF
Windows images cannot be imported. For this reason
`StemcellCloudProperties` now has a new property
`ImageURL` which means that instead of creating a
new GCP image from a tarball (either inside the
stem cell or on GCS), the existing gcp image will
be used for new instances.

[#135362799]

+++ src/bosh-google-cpi/action/delete_stemcell.go
@@ -1,6 +1,8 @@
 package action
 
 import (
+	"strings"
+
 	bosherr "github.com/cloudfoundry/bosh-utils/errors"
 
 	"bosh-google-cpi/google/image_service"
@@ -19,6 +21,10 @@ func NewDeleteStemcell(
 }
 
 func (ds DeleteStemcell) Run(stemcellCID StemcellCID) (interface{}, error) {
+	if strings.HasPrefix(string(stemcellCID), "https://www.googleapis.com/compute/v1/projects/") {
+		return nil, nil
+	}
+
 	if err := ds.imageService.Delete(string(stemcellCID)); err != nil {
 		return nil, bosherr.WrapErrorf(err, "Deleting stemcell '%s'", stemcellCID)
 	}
